Implemented pre-connection JITMs
